# docs/openapi.yaml
openapi: '3.0.2'
info:
  description: >-
    This is the API documentation of Trackpal, a location tracking and alerting solution. You can find the source code
    and/or open a PR [here](https://github.com/dimkouv/trackpal).
  version: 1.0.0
  title: Trackpal
  contact:
    email: dimkouv@protonmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'

servers:
  - url: https://www.trackpal.xyz/api/v1

tags:
  - name: Account
    description: Everything about your user account. You have to create a user account in order to use the API.
  - name: Tracking
    description: Everything about location tracking
  - name: Alerting
    description: Everything about your device alerts

paths:
  /auth/register:
    post:
      tags:
        - Account
      summary: Register a new user account
      description: >-
        An endpoint that allows you to create a new user account. Upon success, you will receive
        a confirmation email. The link will be valid for 10 minutes.
      operationId: authRegister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterRequest'
      responses:
        '201':
          description: Account created
        '400':
          description: Something bad happened
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: [ACCOUNT_EXISTS, INVALID_BODY, INVALID_EMAIL, INSECURE_PASSWORD, GENERIC_ERROR]
        '500':
          description: Internal server error

  /auth/login:
    post:
      tags:
        - Account
      summary: Login
      description: >-
        Get a jwt token that can be used for authenticated requests. This token expires after 5 minutes, in order
        to refresh it before it gets expired you can use the 'Token refresh' endpoint.
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Login success
          content:
            text/plain:
              schema:
                type: string
                example: eyJhbGciOiJIUzI1NiIsInR5...
        '400':
          description: Something bad happened
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: [NOT_ACTIVATED, INVALID_BODY, GENERIC_ERROR]
        '404':
          description: Account with email/password combination not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: [NOT_FOUND]
        '500':
          description: Internal server error

  /auth/refresh:
    post:
      tags:
        - Account
      summary: Token refresh
      description: >-
        When you logged in you received an expiring token. Use this endpoint to refresh your token before it expires.
        This endpoint responds with a new token.
      operationId: authRefresh
      responses:
        '200':
          description: Token was refreshed
          content:
            text/plain:
              schema:
                type: string
                example: eyJhbGciOiJIUzI1NiIsInR5...
        '400':
          description: Something bad happened
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: [TOKEN_EXPIRED, GENERIC_ERROR]
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  /auth/activate:
    post:
      tags:
        - Account
      summary: Activate your account
      description: >-
        After registration you received a token in your email address. Use this token to activate your account.
        Note that this token expires after 10 minutes (TODO: Allow resending a new token).
      operationId: authActivate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthActivateRequest'
      responses:
        '200':
          description: Activation success
        '400':
          description: Something bad happened
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: [INVALID_BODY, TOKEN_EXPIRED, GENERIC_ERROR]
        '404':
          description: Token/Account combination not found or account already active
        '500':
          description: Internal server error

  /tracking/devices:
    get:
      tags:
        - Tracking
      summary: Get all your tracking devices
      description: >-
        Under your user account you have a list of tracking devices. This endpoint fetches a list of their short
        overviews.
      operationId: getTrackingDevices
      responses:
        '200':
          description: Get a list of your tracking devices ascending by date recorded
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrackingDevice'
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

    post:
      tags:
        - Tracking
      summary: Create a new tracking device
      description: >-
        Add a new tracking device under your account
      operationId: postTrackingDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrackingDeviceInput'
      responses:
        '201':
          description: Device created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackingDevice'
        '400':
          description: Something bad happened
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: [INVALID_BODY, INVALID_DATA]
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  /tracking/devices/{deviceID}/records:
    get:
      tags:
        - Tracking
      summary: Get tracking records of device
      description: >-
        Get all the tracking records of a target device
      parameters:
        - name: deviceID
          in: path
          required: true
          description: ID of the target device
          schema:
            type: number
      operationId: getTrackingDeviceRecords
      responses:
        '200':
          description: Get a list of the tracking records of a target device
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TrackingDeviceRecord'
        '400':
          description: Something bad happened
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: [INVALID_VARS]
        '404':
          description: The device was not found (or not owned by you)
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

    post:
      tags:
        - Tracking
      summary: Create a new tracking record
      description: >-
        Create a new tracking record under the target device
      parameters:
        - name: deviceID
          in: path
          required: true
          description: ID of the target device
          schema:
            type: number
      operationId: postTrackingDeviceRecord
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrackingDeviceRecordInput'
      responses:
        '201':
          description: Record created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrackingDeviceRecord'
        '400':
          description: Something bad happened
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: [INVALID_BODY, INVALID_INPUT, INVALID_VARS]
        '404':
          description: The device was not found (or not owned by you)
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  /tracking/devices/{deviceID}/alerting/enable:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrackingDeviceRecordInput'
      tags:
        - Alerting
      summary: "Enable alerting"
      description: >
        Enables alerting for a target device. You have to also
        specify the current position of the device. If the device
        is moved from this position an alert is sent to the user.
      parameters:
        - name: deviceID
          in: path
          required: true
          description: ID of the target device
          schema:
            type: number
      operationId: postAlertingEnable
      responses:
        '200':
          description: Alerting was enabled
        '400':
          description: Something bad happened
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: [INVALID_BODY, INVALID_VARS]
        '404':
          description: The device was not found (or not owned by you)
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

components:
  schemas:
    AuthRegisterRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        firstName:
          type: string
        lastName:
          type: string

    AuthLoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password

    AuthActivateRequest:
      type: object
      required:
        - email
        - token
      properties:
        email:
          type: string
          format: email
        token:
          type: string

    TrackingDevice:
      type: object
      properties:
        createdAt:
          type: string
          format: 'date-time'
        id:
          type: number
        name:
          type: string

    TrackingDeviceRecord:
      type: object
      properties:
        recordedAt:
          type: string
          format: 'date-time'
        location:
          type: object
          properties:
            lat:
              type: number
              format: float
            lng:
              type: number
              format: float

    CreateTrackingDeviceInput:
      type: object
      properties:
        name:
          type: string

    CreateTrackingDeviceRecordInput:
      type: object
      properties:
        recordedAt:
          type: string
          format: 'date-time'
        location:
          type: object
          properties:
            lat:
              type: number
              format: float
            lng:
              type: number
              format: float